WIP: setting get command, moved setting stuff out of profileservice
