+++ src/api/services/ProfileService.ts
@@ -207,54 +207,4 @@ export class ProfileService {
     public async destroy(id: number): Promise<void> {
         await this.profileRepo.destroy(id);
     }
-
-    @validate()
-    public async getSetting(@request(SettingGetRequest) data: any): Promise<Setting> {
-        const profileId = data.params[0];
-        const key = data.params[1];
-
-        // Fetch profile
-        const profileModel = await this.findOne(profileId);
-        const profile = profileModel.toJSON();
-
-        // Iterate over settings and update with matching key
-        for (const setting of profile.Settings) {
-            if (setting.key === key) {
-                return setting;
-            }
-        }
-        throw new NotFoundException(profileId);
-    }
-
-
-    @validate()
-    public async setSetting(@request(SettingUpdateRequest) data: any): Promise<void> {
-        // Create get request
-        const profileId = data.params[0];
-        const key = data.params[1];
-        const settingGetRequest = {
-            profileId,
-            key
-        } as SettingGetRequest;
-
-        // Update setting
-        const setting = await this.getSetting(settingGetRequest);
-        await this.settingService.update(setting.id, data);
-    }
-
-
-    @validate()
-    public async removeSetting(@request(SettingRemoveRequest) data: any): Promise<void> {
-        // Create get request
-        const profileId = data.params[0];
-        const key = data.params[1];
-        const settingGetRequest = {
-            profileId,
-            key
-        } as SettingGetRequest;
-
-        // Remove setting
-        const setting = await this.getSetting(settingGetRequest);
-        await this.settingService.destroy(setting.id);
-    }
 }
